feat: added exceptions in user methods

+++ src/modules/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpStatus, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { PrismaService } from 'src/prisma/prisma.service';
@@ -20,26 +20,37 @@ export class UserService {
     const user = await this.prisma.user.findUnique({ where: { id }, include: { activities: true } });
 
     if (user) {
-      return {...user, password: undefined}
+      return await { ...user, password: undefined }
     }
 
     throw new NotFoundException("user not found :(");
   }
 
   async update(id: UUID, updateUserDto: UpdateUserDto) {
-    return this.prisma.user.update({ where: { id }, data: updateUserDto })
+    try {
+      return this.prisma.user.update({ where: { id }, data: updateUserDto })
+    } catch (error) {
+      error.status = HttpStatus.BAD_REQUEST;
+    }
   }
 
-  remove(id: UUID) {
-    return this.prisma.user.delete({ where: { id } });
+  async remove(id: UUID) {
+    const user = await this.prisma.user.findUnique({ where: { id } });
+
+    if (user) {
+      return this.prisma.user.delete({ where: { id } });
+    } else {
+      throw new NotFoundException("user not found :(");
+    }
   }
 
   async findByEmail(email: string) {
-    if (!email) {
-      return null;
+    const user = await this.prisma.user.findUnique({ where: { email }, include: { activities: true } });
+
+    if (user) {
+      return  {...user }
+    } else {
+      return new NotFoundException("user not found :(");
     }
-
-    return await this.prisma.user.findUnique({ where: { email }, include: { activities: true } });
-
   }
 }
